pyproject.toml
[project]
name = "vectorless-rag-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=2.32.0",
    "pageindex>=0.2.8",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
]
